+++ src/eastwood/linters/unused.clj
@@ -418,4 +418,72 @@ discarded inside null: null'."
   (apply unused-ret-vals-2 :inside-try args))
 
 
+;; Unused metadata on macro invocations (depends upon macro
+;; definition, but most macros ignore it)
+
+(defn unused-meta-on-macro [{:keys [asts]}]
+  (let [macro-invokes (->> asts
+                           (mapcat ast/nodes)
+                           (filter #(and (:raw-forms %)
+                                         (-> % :raw-forms first meta))))]
+    (for [ast macro-invokes
+          :let [orig-form (-> ast :raw-forms first)
+                loc (-> orig-form meta)
+                non-loc-meta-keys (-> (set (keys loc))
+                                      (set/difference #{:file :line :column
+                                                        :end-line :end-column}))
+                resolved-form (-> ast :eastwood/partly-resolved-forms first)
+                resolved-macro-symbol (try
+                                        (first resolved-form)
+                                        (catch Exception e
+;;                                          (println (format "eastwood-dbg: resolved-form=%s ast="
+;;                                                           resolved-form))
+;;                                          (util/pprint-ast-node ast)
+                                          nil))
+                removed-meta-keys
+                (cond (= :new (:op ast)) non-loc-meta-keys
+
+                      (#{:instance-call :static-call
+                         :instance-field :static-field :host-interop} (:op ast))
+                      (disj non-loc-meta-keys :tag)
+
+                      ;; No metadata is removed for invocations of
+                      ;; clojure.core/fn -- It is implemented
+                      ;; specially to do this by examining the
+                      ;; metadata of its &form argument.
+                      (= resolved-macro-symbol 'clojure.core/fn) []
+
+                      :else non-loc-meta-keys)
+
+                sorted-removed-meta-keys (sort removed-meta-keys)
+                ]
+          :when (seq removed-meta-keys)]
+
+      (util/add-loc-info loc
+       {:linter :unused-meta-on-macro
+        :msg
+        (cond
+         (= :new (:op ast))
+         (format "Java constructor call '%s' has metadata with keys %s.  All metadata is eliminated from such forms during macroexpansion and thus ignored by Clojure."
+                 (first orig-form)
+                 sorted-removed-meta-keys)
+         
+         (#{:instance-call :static-call :instance-field :static-field
+            :host-interop} (:op ast))
+         (format "Java %s '%s' has metadata with keys %s.  All metadata keys except :tag are eliminated from such forms during macroexpansion and thus ignored by Clojure."
+                 (case (:op ast)
+                   :instance-call "instance method call"
+                   :static-call "static method call"
+                   :instance-field "instance field access"
+                   :static-field "static field access"
+                   :host-interop "instance method/field access")
+                 (first orig-form)
+                 sorted-removed-meta-keys)
+         
+         :else
+         (format "Macro invocation of '%s' has metadata with keys %s that are almost certainly ignored."
+                 (first orig-form)
+                 sorted-removed-meta-keys))}))))
+
+
 ;; TODO: Unused locals